requirements.txt
streamlit>=1.25.0
pandas>=1.3.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
nbimporter>=0.3.4
google-cloud-documentai>=2.13.0
google-cloud-storage>=2.0.0
ipykernel>=6.0.0 